Temporary workaround for generating unique model names (#47)
If two schema different schemas share the same name, fastapi openapi
documentation fails.

This is a temporary work-around for openapi spec generation purposes. It
will work only for top level model names.

We've bumped into this issue while hosting 2 runnables that start with a
PromptTemplate (same name for the schema), but have different template
variables (so the schema are different).

We'll need to find a better solution to handle nested models. This
should be good for now as it'll allow users to deploy the common case
correctly.

---------

Co-authored-by: Erick Friis <erickfriis@gmail.com>