Clean up endpoints (#338)
Remove feedback related endpoints that take a config hash. The feedback
endpoints only care about the run_id and the feedback, so a config hash
should not be needed.

Also removing the `get` version of checking if feedback is enabled. I'll
transition both the UI and the backend in a bit.