add: Support for Server-Sent Events in APIHandler
This commit adds support for Server-Sent Events (SSE) in the APIHandler class. It imports the necessary modules from `sse_starlette` and includes a check for the `ServerSentEvent` class. If SSE is supported and the chunk is an instance of `ServerSentEvent`, it yields the chunk. This change improves the functionality of the APIHandler by enabling SSE support.